adjusted build scripts to use new libs directory

+++ buildLEDFun.sh
@@ -5,10 +5,14 @@
 #
 
 # Build ledfun app
-rm main.swift
+
+if [ -e "main.swift" ]
+then
+    rm main.swift
+fi
 
 ln -s LEDFun.swift main.swift
 
-swiftc -v -o ledfun -I lib/SMBus-swift/Packages/Ci2c -I lib/SMBus-swift/Packages/CioctlHelper lib/SMBus-swift/Sources/SMBus.swift Signals.swift LEDBackpack.swift main.swift
+swiftc -v -o ledfun -I libs/SMBus-swift/Packages/Ci2c -I libs/SMBus-swift/Packages/CioctlHelper libs/SMBus-swift/Sources/SMBus.swift Signals.swift LEDBackpack.swift main.swift
 
 rm main.swift
