+++ buildPi.sh
@@ -3,10 +3,15 @@
 # Since the Raspberry Pi does not currently have a working Package Manager, we need
 # to pull it all together into a compile script
 #
-rm main.swift
+
+if [ -e "main.swift" ]
+then
+    rm main.swift
+fi
+
 # Build the pi app
 ln -s Pi.swift main.swift
 
-swiftc -v -o pi -I lib/SMBus-swift/Packages/Ci2c -I lib/SMBus-swift/Packages/CioctlHelper lib/SMBus-swift/Sources/SMBus.swift Signals.swift lib/SipHash/sources/*.swift lib/BigInt/sources/*.swift Coroutine.swift LEDBackpack.swift main.swift
+swiftc -v -o pi -I libs/SMBus-swift/Packages/Ci2c -I libs/SMBus-swift/Packages/CioctlHelper libs/SMBus-swift/Sources/SMBus.swift Signals.swift libs/SipHash/sources/*.swift libs/BigInt/sources/*.swift Coroutine.swift LEDBackpack.swift main.swift
 
 rm main.swift
